[Fixes #7] Work around Pinboard.in invalid XML.
Pinboard.in is returning invalid XML from `posts/update`. This works around the issue by checking for the existence of responseXML before looking for the <update> tag.

I also added a fallback method which searches for the update time in responseText (i.e. the invalid XML string). It seems to be working quite well.